refactor: remove claims to entity
